+++ sdk/identity/identity/src/credentials/chainedTokenCredential.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import { AccessToken, TokenCredential, GetTokenOptions } from "@azure/core-http";
-import { AggregateAuthenticationError, CredentialUnavailableError } from "../client/errors";
+import { AggregateAuthenticationError, CredentialUnavailable } from "../client/errors";
 import { createSpan } from "../util/tracing";
 import { SpanStatusCode } from "@azure/core-tracing";
 import { credentialLogger, formatSuccess, formatError } from "../util/logging";
@@ -54,7 +54,10 @@ export class ChainedTokenCredential implements TokenCredential {
    * @param options - The options used to configure any requests this
    *                `TokenCredential` implementation might make.
    */
-  async getToken(scopes: string | string[], options?: GetTokenOptions): Promise<AccessToken> {
+  async getToken(
+    scopes: string | string[],
+    options?: GetTokenOptions
+  ): Promise<AccessToken | null> {
     let token = null;
     let successfulCredentialName = "";
     const errors = [];
@@ -93,7 +96,7 @@ export class ChainedTokenCredential implements TokenCredential {
     logger.getToken.info(`Result for ${successfulCredentialName}: ${formatSuccess(scopes)}`);
 
     if (token === null) {
-      throw new CredentialUnavailableError("Failed to retrieve a valid token");
+      throw new CredentialUnavailable("Failed to retrieve a valid token");
     }
     return token;
   }