+++ sdk/identity/identity/src/credentials/interactiveBrowserCredential.ts
@@ -6,13 +6,9 @@
 import { AccessToken, GetTokenOptions, TokenCredential } from "@azure/core-http";
 import { credentialLogger } from "../util/logging";
 import { trace } from "../util/tracing";
-import { AuthenticationRecord } from "../msal/types";
 import { MsalOpenBrowser } from "../msal/nodeFlows/msalOpenBrowser";
 import { MsalFlow } from "../msal/flows";
-import {
-  InteractiveBrowserCredentialBrowserOptions,
-  InteractiveBrowserCredentialOptions
-} from "./interactiveBrowserCredentialOptions";
+import { InteractiveBrowserCredentialOptions } from "./interactiveBrowserCredentialOptions";
 
 const logger = credentialLogger("InteractiveBrowserCredential");
 
@@ -27,16 +23,13 @@ const logger = credentialLogger("InteractiveBrowserCredential");
  */
 export class InteractiveBrowserCredential implements TokenCredential {
   private msalFlow: MsalFlow;
-  private disableAutomaticAuthentication?: boolean;
 
   /**
    * Creates an instance of InteractiveBrowserCredential with the details needed.
    *
    * @param options - Options for configuring the client which makes the authentication requests.
    */
-  constructor(
-    options: InteractiveBrowserCredentialOptions | InteractiveBrowserCredentialBrowserOptions = {}
-  ) {
+  constructor(options: InteractiveBrowserCredentialOptions = {}) {
     const redirectUri =
       typeof options.redirectUri === "function"
         ? options.redirectUri()
@@ -48,7 +41,6 @@ export class InteractiveBrowserCredential implements TokenCredential {
       logger,
       redirectUri
     });
-    this.disableAutomaticAuthentication = options?.disableAutomaticAuthentication;
   }
 
   /**
@@ -68,33 +60,7 @@ export class InteractiveBrowserCredential implements TokenCredential {
   async getToken(scopes: string | string[], options: GetTokenOptions = {}): Promise<AccessToken> {
     return trace(`${this.constructor.name}.getToken`, options, async (newOptions) => {
       const arrayScopes = Array.isArray(scopes) ? scopes : [scopes];
-      return this.msalFlow.getToken(arrayScopes, {
-        ...newOptions,
-        disableAutomaticAuthentication: this.disableAutomaticAuthentication
-      });
-    });
-  }
-
-  /**
-   * Authenticates with Azure Active Directory and returns an access token if
-   * successful.  If authentication cannot be performed at this time, this method may
-   * return null.  If an error occurs during authentication, an {@link AuthenticationError}
-   * containing failure details will be thrown.
-   *
-   * If the token can't be retrieved silently, this method will require user interaction to retrieve the token.
-   *
-   * @param scopes - The list of scopes for which the token will have access.
-   * @param options - The options used to configure any requests this
-   *                  TokenCredential implementation might make.
-   */
-  async authenticate(
-    scopes: string | string[],
-    options: GetTokenOptions = {}
-  ): Promise<AuthenticationRecord | undefined> {
-    return trace(`${this.constructor.name}.authenticate`, options, async (newOptions) => {
-      const arrayScopes = Array.isArray(scopes) ? scopes : [scopes];
-      await this.msalFlow.getToken(arrayScopes, newOptions);
-      return this.msalFlow.getActiveAccount();
+      return this.msalFlow.getToken(arrayScopes, newOptions);
     });
   }
 }