+++ sdk/identity/identity/src/index.ts
@@ -6,10 +6,7 @@ import { DefaultAzureCredential } from "./credentials/defaultAzureCredential";
 
 export { AuthenticationRecord } from "./msal/types";
 export { AuthenticationRequiredError } from "./msal/errors";
-export { serializeAuthenticationRecord, deserializeAuthenticationRecord } from "./msal/utils";
 export { TokenCredentialOptions } from "./client/identityClient";
-export { InteractiveCredentialOptions } from "./credentials/interactiveCredentialOptions";
-export { TokenCachePersistenceOptions } from "./tokenCache/persistencePlatforms";
 
 export { ChainedTokenCredential } from "./credentials/chainedTokenCredential";
 export {
@@ -29,7 +26,6 @@ export { AzureCliCredential } from "./credentials/azureCliCredential";
 export { InteractiveBrowserCredential } from "./credentials/interactiveBrowserCredential";
 export {
   InteractiveBrowserCredentialOptions,
-  InteractiveBrowserCredentialBrowserOptions,
   BrowserLoginStyle,
   InteractiveBrowserAuthenticationFlow
 } from "./credentials/interactiveBrowserCredentialOptions";
@@ -54,7 +50,7 @@ export {
   AggregateAuthenticationError,
   AuthenticationErrorName,
   AggregateAuthenticationErrorName,
-  CredentialUnavailableError,
+  CredentialUnavailable,
   CredentialUnavailableErrorName
 } from "./client/errors";