+++ sdk/identity/identity/src/msal/browserFlows/browserCommon.ts
@@ -12,7 +12,7 @@ import { MsalFlow, MsalFlowOptions } from "../flows";
 import { AuthenticationRecord } from "../types";
 import { CredentialFlowGetTokenOptions } from "../credentials";
 import { AuthenticationRequiredError } from "../errors";
-import { CredentialUnavailableError } from "../../client/errors";
+import { CredentialUnavailable } from "../../client/errors";
 
 /**
  * Union of the constructor parameters that all MSAL flow types take.
@@ -77,7 +77,7 @@ export abstract class MsalBrowser extends MsalBaseUtilities implements MsalBrows
     this.logger = options.logger;
     this.loginStyle = options.loginStyle;
     if (!options.clientId) {
-      throw new CredentialUnavailableError("A client ID is required in browsers");
+      throw new CredentialUnavailable("A client ID is required in browsers");
     }
     this.clientId = options.clientId;
     this.tenantId = resolveTenantId(this.logger, options.tenantId, options.clientId);