+++ sdk/identity/identity/test/internal/node/clientCertificateCredential.spec.ts
@@ -8,9 +8,8 @@ import assert from "assert";
 import * as path from "path";
 import { env, isPlaybackMode } from "@azure/test-utils-recorder";
 import { ConfidentialClientApplication } from "@azure/msal-node";
-import { ClientCertificateCredential, TokenCachePersistenceOptions } from "../../../src";
+import { ClientCertificateCredential } from "../../../src";
 import { MsalTestCleanup, msalNodeTestSetup } from "../../msalTestUtils";
-import { TokenCachePersistence } from "../../../src/tokenCache/TokenCachePersistence";
 import { MsalNode } from "../../../src/msal/nodeFlows/nodeCommon";
 import { Context } from "mocha";
 
@@ -85,82 +84,5 @@ describe("ClientCertificateCredential (internal)", function() {
     } catch (e) {
       return;
     }
-
-    it("Accepts tokenCachePersistenceOptions", async function(this: Context) {
-      if (isPlaybackMode()) {
-        // MSAL creates a client assertion based on the certificate that I haven't been able to mock.
-        // This assertion could be provided as parameters, but we don't have that in the public API yet,
-        // and I'm trying to avoid having to generate one ourselves.
-        this.skip();
-      }
-      // OSX asks for passwords on CI, so we need to skip these tests from our automation
-      if (process.platform === "darwin") {
-        this.skip();
-      }
-
-      const tokenCachePersistenceOptions: TokenCachePersistenceOptions = {
-        name: this.test?.title.replace(/[^a-zA-Z]/g, "_"),
-        allowUnencryptedStorage: true
-      };
-
-      // Emptying the token cache before we start.
-      const tokenCache = new TokenCachePersistence(tokenCachePersistenceOptions);
-      const persistence = await tokenCache.getPersistence();
-      persistence?.save("{}");
-
-      const credential = new ClientCertificateCredential(
-        env.AZURE_TENANT_ID,
-        env.AZURE_CLIENT_ID,
-        certificatePath,
-        { tokenCachePersistenceOptions }
-      );
-
-      await credential.getToken(scope);
-      const result = await persistence?.load();
-      const parsedResult = JSON.parse(result!);
-      assert.ok(parsedResult.AccessToken);
-    });
-
-    it("Authenticates silently with tokenCachePersistenceOptions", async function(this: Context) {
-      if (isPlaybackMode()) {
-        // MSAL creates a client assertion based on the certificate that I haven't been able to mock.
-        // This assertion could be provided as parameters, but we don't have that in the public API yet,
-        // and I'm trying to avoid having to generate one ourselves.
-        this.skip();
-      }
-      // OSX asks for passwords on CI, so we need to skip these tests from our automation
-      if (process.platform === "darwin") {
-        this.skip();
-      }
-
-      const tokenCachePersistenceOptions: TokenCachePersistenceOptions = {
-        name: this.test?.title.replace(/[^a-zA-Z]/g, "_"),
-        allowUnencryptedStorage: true
-      };
-
-      // Emptying the token cache before we start.
-      const tokenCache = new TokenCachePersistence(tokenCachePersistenceOptions);
-      const persistence = await tokenCache.getPersistence();
-      persistence?.save("{}");
-
-      const credential = new ClientCertificateCredential(
-        env.AZURE_TENANT_ID,
-        env.AZURE_CLIENT_ID,
-        certificatePath,
-        { tokenCachePersistenceOptions }
-      );
-
-      await credential.getToken(scope);
-      assert.equal(getTokenSilentSpy.callCount, 1);
-      assert.equal(doGetTokenSpy.callCount, 1);
-
-      await credential.getToken(scope);
-      assert.equal(getTokenSilentSpy.callCount, 2);
-      // Even though we're providing a file persistence cache,
-      // The Client Credential flow does not return the account information from the authentication service,
-      // so each time getToken gets called, we will have to acquire a new token through the service.
-      // MSAL also doesn't store the account in the cache (getAllAccounts returns an empty array).
-      assert.equal(doGetTokenSpy.callCount, 2);
-    });
   });
 });