+++ sdk/identity/identity/test/internal/node/interactiveBrowserCredential.spec.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 /* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */
+/* eslint-disable @azure/azure-sdk/ts-no-namespaces */
 
 import Sinon from "sinon";
 import assert from "assert";
@@ -10,7 +11,15 @@ import { env } from "@azure/test-utils-recorder";
 import { InteractiveBrowserCredential } from "../../../src";
 import { MsalTestCleanup, msalNodeTestSetup } from "../../msalTestUtils";
 import { interactiveBrowserMockable } from "../../../src/msal/nodeFlows/msalOpenBrowser";
-import { isNode8 } from "../../../src/tokenCache/requireMsalNodeExtensions";
+import { URL } from "url";
+
+declare global {
+  namespace NodeJS {
+    interface Global {
+      URL: typeof import("url").URL;
+    }
+  }
+}
 
 describe("InteractiveBrowserCredential (internal)", function() {
   let cleanup: MsalTestCleanup;
@@ -29,8 +38,8 @@ describe("InteractiveBrowserCredential (internal)", function() {
 
   it("Throws an expected error if no browser is available", async function(this: Context) {
     // On Node 8, URL is not defined. We use URL on the msalOpenBrowser.ts file.
-    if (isNode8) {
-      this.skip();
+    if (process.version.startsWith("v8.")) {
+      global.URL = URL;
     }
 
     // The SinonStub type does not include this second parameter to throws().